fix power issue with jax.jacfwd, for some reason making m float made the trick

+++ desc/basis.py
@@ -1510,7 +1510,7 @@ def zernike_radial(r, l, m, dr=0):
         basis function(s) evaluated at specified points
 
     """
-    m = jnp.abs(m)
+    m = jnp.abs(m).astype(float)
     alpha = m
     beta = 0
     n = (l - m) // 2
